Refactor and new feature of dailyclean.py
The script has been refactored with additional functions, making it more readable and easier to maintain.

The `label_selector` for `stateful_set` has been modified. It now uses `axa.com/dailyclean!=false` to be consistent with the `deployment label_selector`.

Added functionality to handle the metadata label `axa.com/dailyclean-start-replicas`. If this label is set on a deployment or stateful_set, its value will be utilized when the start script is executed.